Push minimal duplication to support GLES and OpenGL in one driver
Use the gles tag to set gles for testing.
Platform auto-switching is next
