+++ internal/driver/gl/gl_es.go
@@ -1,0 +1,225 @@
+// +build gles
+
+package gl
+
+import (
+	"fmt"
+	"image"
+	"image/draw"
+	"strings"
+
+	gl "github.com/go-gl/gl/v3.1/gles2"
+	"github.com/go-gl/glfw/v3.2/glfw"
+
+	"fyne.io/fyne"
+	"fyne.io/fyne/theme"
+)
+
+func newTexture() uint32 {
+	var texture uint32
+
+	gl.GenTextures(1, &texture)
+	gl.ActiveTexture(gl.TEXTURE0)
+	gl.BindTexture(gl.TEXTURE_2D, texture)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
+	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
+
+	return texture
+}
+
+func (c *glCanvas) imgToTexture(img image.Image) uint32 {
+	switch i := img.(type) {
+	case *image.Uniform:
+		texture := newTexture()
+		r, g, b, a := i.RGBA()
+		r8, g8, b8, a8 := uint8(r>>8), uint8(g>>8), uint8(b>>8), uint8(a>>8)
+		data := []uint8{r8, g8, b8, a8}
+		gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, 1, 1, 0, gl.RGBA,
+			gl.UNSIGNED_BYTE, gl.Ptr(data))
+		return texture
+	case *image.RGBA:
+		if len(i.Pix) == 0 { // image is empty
+			return 0
+		}
+
+		var texture uint32
+		texture = newTexture()
+		gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, int32(i.Rect.Size().X), int32(i.Rect.Size().Y),
+			0, gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(i.Pix))
+		return texture
+	default:
+		rgba := image.NewRGBA(image.Rect(0, 0, img.Bounds().Dx(), img.Bounds().Dy()))
+		draw.Draw(rgba, rgba.Rect, img, image.ZP, draw.Over)
+		return c.imgToTexture(rgba)
+	}
+}
+
+func setViewport(width, height int) {
+	gl.Viewport(0, 0, int32(width), int32(height))
+}
+
+func freeTexture(obj fyne.CanvasObject) {
+	texture := textures[obj]
+	if texture != 0 {
+		gl.DeleteTextures(1, &texture)
+		delete(textures, obj)
+	}
+}
+
+func initWindowHints() {
+	glfw.WindowHint(glfw.ClientAPI, glfw.OpenGLESAPI)
+	glfw.WindowHint(glfw.ContextVersionMajor, 2)
+	glfw.WindowHint(glfw.ContextVersionMinor, 0)
+}
+
+func glInit() {
+	err := gl.Init()
+	if err != nil {
+		fyne.LogError("failed to initialise OpenGL", err)
+		return
+	}
+
+	gl.Disable(gl.DEPTH_TEST)
+}
+
+func compileShader(source string, shaderType uint32) (uint32, error) {
+	shader := gl.CreateShader(shaderType)
+
+	csources, free := gl.Strs(source)
+	gl.ShaderSource(shader, 1, csources, nil)
+	free()
+	gl.CompileShader(shader)
+
+	var status int32
+	gl.GetShaderiv(shader, gl.COMPILE_STATUS, &status)
+	if status == gl.FALSE {
+		var logLength int32
+		gl.GetShaderiv(shader, gl.INFO_LOG_LENGTH, &logLength)
+
+		info := strings.Repeat("\x00", int(logLength+1))
+		gl.GetShaderInfoLog(shader, logLength, nil, gl.Str(info))
+
+		return 0, fmt.Errorf("failed to compile %v: %v", source, info)
+	}
+
+	return shader, nil
+}
+
+const (
+	vertexShaderSource = `
+    #version 100
+    attribute vec3 vert;
+    attribute vec2 vertTexCoord;
+    varying highp vec2 fragTexCoord;
+
+    void main() {
+        fragTexCoord = vertTexCoord;
+
+        gl_Position = vec4(vert, 1);
+    }
+` + "\x00"
+
+	fragmentShaderSource = `
+    #version 100
+    uniform sampler2D tex;
+
+    varying highp vec2 fragTexCoord;
+
+    void main() {
+        gl_FragColor = texture2D(tex, fragTexCoord);
+    }
+` + "\x00"
+)
+
+func (c *glCanvas) initOpenGL() {
+	vertexShader, err := compileShader(vertexShaderSource, gl.VERTEX_SHADER)
+	if err != nil {
+		panic(err)
+	}
+	fragmentShader, err := compileShader(fragmentShaderSource, gl.FRAGMENT_SHADER)
+	if err != nil {
+		panic(err)
+	}
+
+	prog := gl.CreateProgram()
+	gl.AttachShader(prog, vertexShader)
+	gl.AttachShader(prog, fragmentShader)
+	gl.LinkProgram(prog)
+
+	c.program = prog
+}
+
+func (c *glCanvas) glClearBuffer() {
+	gl.UseProgram(c.program)
+
+	r, g, b, a := theme.BackgroundColor().RGBA()
+	max16bit := float32(255 * 255)
+	gl.ClearColor(float32(r)/max16bit, float32(g)/max16bit, float32(b)/max16bit, float32(a)/max16bit)
+	gl.Clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT)
+}
+
+func (c *glCanvas) glScissorOpen(x, y, w, h int32) {
+	gl.Scissor(x, y, w, h)
+	gl.Enable(gl.SCISSOR_TEST)
+}
+
+func (c *glCanvas) glScissorClose() {
+	gl.Disable(gl.SCISSOR_TEST)
+}
+
+func (c *glCanvas) glCreateBuffer(points []float32) (uint32, uint32) {
+	var vao uint32
+	gl.GenVertexArrays(1, &vao)
+	gl.BindVertexArray(vao)
+	gl.EnableVertexAttribArray(0)
+
+	var vbo uint32
+	gl.GenBuffers(1, &vbo)
+	gl.BindBuffer(gl.ARRAY_BUFFER, vbo)
+	gl.BufferData(gl.ARRAY_BUFFER, 4*len(points), gl.Ptr(points), gl.STATIC_DRAW)
+
+	textureUniform := gl.GetUniformLocation(c.program, gl.Str("tex\x00"))
+	gl.Uniform1i(textureUniform, 0)
+
+	vertAttrib := uint32(gl.GetAttribLocation(c.program, gl.Str("vert\x00")))
+	gl.EnableVertexAttribArray(vertAttrib)
+	gl.VertexAttribPointer(vertAttrib, 3, gl.FLOAT, false, 5*4, gl.PtrOffset(0))
+
+	texCoordAttrib := uint32(gl.GetAttribLocation(c.program, gl.Str("vertTexCoord\x00")))
+	gl.EnableVertexAttribArray(texCoordAttrib)
+	gl.VertexAttribPointer(texCoordAttrib, 2, gl.FLOAT, false, 5*4, gl.PtrOffset(3*4))
+
+	return vao, vbo
+}
+
+func (c *glCanvas) glFreeBuffer(vao, vbo uint32) {
+	gl.BindVertexArray(0)
+	gl.DeleteVertexArrays(1, &vao)
+
+	gl.BindBuffer(gl.ARRAY_BUFFER, 0)
+	gl.DeleteBuffers(1, &vbo)
+}
+
+func (c *glCanvas) glDrawTexture(texture uint32, points []float32, alpha float32) {
+	gl.Enable(gl.BLEND)
+	// here we have to choose between blending the image alpha or fading it...
+	// TODO find a way to support both
+	if alpha != 1.0 {
+		gl.BlendColor(0, 0, 0, alpha)
+		gl.BlendFunc(gl.CONSTANT_ALPHA, gl.ONE_MINUS_CONSTANT_ALPHA)
+	} else {
+		gl.BlendFunc(gl.ONE, gl.ONE_MINUS_SRC_ALPHA)
+	}
+
+	gl.ActiveTexture(gl.TEXTURE0)
+	gl.BindTexture(gl.TEXTURE_2D, texture)
+
+	gl.DrawArrays(gl.TRIANGLE_STRIP, 0, int32(len(points)/5))
+}
+
+func (c *glCanvas) glCapture(width, height int32, pixels *[]uint8) {
+	gl.ReadBuffer(gl.FRONT)
+	gl.ReadPixels(0, 0, int32(width), int32(height), gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(*pixels))
+}
