+++ internal/driver/gl/gl_other.go
@@ -3,8 +3,6 @@
 
 package gl
 
-import gl "github.com/go-gl/gl/v3.1/gles2"
-
 func updateGLContext(w *window) {
 	canvas := w.Canvas()
 	size := canvas.Size()
@@ -12,7 +10,7 @@ func updateGLContext(w *window) {
 	winWidth := scaleInt(canvas, size.Width)
 	winHeight := scaleInt(canvas, size.Height)
 
-	gl.Viewport(0, 0, int32(winWidth), int32(winHeight))
+	setViewport(winWidth, winHeight)
 }
 
 func forceWindowRefresh(_ *window) {
