+++ internal/driver/gl/gl_windows.go
@@ -1,7 +1,5 @@
 package gl
 
-import gl "github.com/go-gl/gl/v3.1/gles2"
-
 func updateGLContext(w *window) {
 	canvas := w.Canvas()
 	size := canvas.Size()
@@ -9,7 +7,7 @@ func updateGLContext(w *window) {
 	winWidth := scaleInt(canvas, size.Width)
 	winHeight := scaleInt(canvas, size.Height)
 
-	gl.Viewport(0, 0, int32(winWidth), int32(winHeight))
+	setViewport(winWidth, winHeight)
 }
 
 // This forces a redraw of the window as we resize or move
